LRCI-3823 Create more base github event handlers to move logic into correct locations
